11_python_api/requirements.txt
# RhythmIQ Python API Requirements
flask>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
pillow>=8.0.0
joblib>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0